requirements.txt
google-generativeai
tokenizers
transformers
accelerate
flash-attn
einops
omegaconf
flask
flask_cors
flask_socketio
eventlet
gunicorn
ctranslate2 # for cuda 12
opencv-python
basicsr
kornia
face-alignment==1.3.4
facexlib
librosa
dlib
numpy==1.23.4
trainer
pypinyin
hangul_romanize
num2words
spacy
pandas
pysbd
librosa==0.9.2
python-dotenv